+++ src/domains/opinions/team-opinions-comments/team-opinions-comments.repository.ts
@@ -4,11 +4,13 @@ import { FirebaseService } from 'src/database/firebase/firebase.service';
 @Injectable()
 export class TeamOpinionsCommentsRepository {
     private teamOpinionsCollection: FirebaseFirestore.CollectionReference;
+    private teamCommentsCollection: FirebaseFirestore.CollectionReference;
 
     constructor(private firebaseService: FirebaseService) {}
 
     async onModuleInit() {
         this.teamOpinionsCollection = this.firebaseService.teamOpinionsCollection;
+        this.teamCommentsCollection = this.firebaseService.playerCommentsCollection;
     }
 
     // -------------------------- Team Opinions --------------------------//
@@ -51,32 +53,45 @@ export class TeamOpinionsCommentsRepository {
 
     // -------------------------- Team Comments --------------------------//
 
-    async addComment(opinionId: string, comment: any): Promise<any> {
+    async addComment(comment: any): Promise<any> {
         const plainComment = JSON.parse(JSON.stringify(comment));
-        const commentsRef = this.teamOpinionsCollection.doc(opinionId).collection('comments');
-        return commentsRef.add(plainComment);
+        const commentsRef = await this.teamCommentsCollection.add(plainComment);
+        return { id: commentsRef.id, ...comment };
     }
 
     async getCommentsForOpinion(opinionId: string): Promise<any[]> {
-        const snapshot = await this.teamOpinionsCollection.doc(opinionId).collection('comments').get();
+        const snapshot = await this.teamCommentsCollection.where('opinion_team_id', '==', opinionId).get();
         return snapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
     }
 
     async getCommentByIdForOpinion(opinionId: string, commentId: string): Promise<any> {
-        const commentRef = this.teamOpinionsCollection.doc(opinionId).collection('comments').doc(commentId);
+        const commentRef = this.teamCommentsCollection.doc(commentId);
         const commentSnapshot = await commentRef.get();
+        const commentData = commentSnapshot.data();
+        if (!commentData || commentData.opinion_team_id !== opinionId) {
+            return null;
+        }
         return commentSnapshot.exists ? { id: commentSnapshot.id, ...commentSnapshot.data() } : null;
     }
 
     async updateComment(opinionId: string, commentId: string, comment: any): Promise<any> {
         const plainComment = JSON.parse(JSON.stringify(comment));
-        await this.teamOpinionsCollection.doc(opinionId).collection('comments').doc(commentId).update(plainComment);
-        return { id: commentId, ...comment };
+        const commentRef = this.teamCommentsCollection.doc(commentId);
+        const commentSnapshot = await commentRef.get();
+        const commentData = commentSnapshot.data();
+        if (!commentData || commentData.opinion_team_id !== opinionId) {
+            return null;
+        }
+        await commentRef.update(plainComment);
     }
 
     async deleteComment(opinionId: string, commentId: string): Promise<any> {
-        const commentRef = this.teamOpinionsCollection.doc(opinionId).collection('comments').doc(commentId);
+        const commentRef = this.teamCommentsCollection.doc(commentId);
         const commentSnapshot = await commentRef.get();
+        const commentData = commentSnapshot.data();
+        if (!commentData || commentData.opinion_team_id !== opinionId) {
+            return null;
+        }
         const deletedID = commentSnapshot.id;
         const deletedData = commentSnapshot.data();
         await commentRef.delete();
